requirements.txt
# GraphRAG - Turkish Legal System

networkx>=3.0
numpy>=1.24
faiss-cpu>=1.7
python-dotenv>=1.0
sentence-transformers>=2.2

# API
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0

# CLI
textual>=0.50

# Optional: only required if using OpenAI embeddings
# openai>=1.0
